requirements.txt
django==2.1.13
django-redis
djangorestframework
django-filter
django-cors-headers
django-imagekit
django-debug-toolbar
PyJWT
qiniu
qrcode
openpyxl
xlrd
xlwt
requests
celery
itsdangerous
elasticsearch
PyMySQL
pillow